Change function js removeFriend

+++ src/Services/FriendshipManager.php
@@ -56,8 +56,12 @@ class FriendshipManager
     {
         $friendship = $this->get($user, $friend);
 
+        if($friendship == null) return false;
+
         $this->entityManager->remove($friendship);
         $this->entityManager->flush();
+
+        return true;
     }
 
     /**
